Apply code review fixes from PR #5 analysis (#6)
Agent-Logs-Url: https://github.com/singer-io/tap-clubspeed/sessions/49b19b4f-e77d-4e4c-ac87-d0fa1c82f593
