+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.2
+  * Upgrade Python runtime from 3.9 to 3.12 [#5](https://github.com/singer-io/tap-clubspeed/pull/5)
+  * Bump `singer-python` from 5.14.3 to 6.8.0 [#5](https://github.com/singer-io/tap-clubspeed/pull/5)
+  * Bump `requests` from 2.32.5 to 2.33.0 [#5](https://github.com/singer-io/tap-clubspeed/pull/5)
+  * Add unit and integration test suite [#5](https://github.com/singer-io/tap-clubspeed/pull/5)
+
 ## 1.0.1
   * Bump dependencies [#4](https://github.com/singer-io/tap-clubspeed/pull/4)
